tests: Move etcdctl to e2e framework
